pyproject.toml
[project]
name = "agent-template"
version = "0.1.0"
description = "A template for A2A agents"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.20",
    "anthropic>=0.34.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "openai>=1.0.0",
    "together>=1.2.0",
    "uvicorn>=0.38.0",
]
